feat(spa): add EcoDataComponent, the app-wide component base
Introduces EcoData.Spa.Blazor: base classes every EcoData component renders
from, carrying the cross-cutting concerns each app would otherwise wire by
hand. Localization is the first one.

EcoDataComponent is a Tempest StatefulComponent that injects ILocalizer as L,
subscribes LanguageChanged once, and exposes a virtual OnLanguageChanged hook
for work beyond redrawing — re-titling the navbar, refetching locale-dependent
data. The re-render happens either way, through Tempest's Mutate primitive so
it marshals and costs exactly one pass. EcoDataLayout mirrors it for layouts;
C# has no multiple inheritance, so the two are twins, the same split Tempest
makes for the same reason.

The library takes no MudBlazor: component libraries build on these bases, and
a host that wants a design system brings its own.

FaunaFinder's LocalizedComponentBase is deleted and its 27 consumers now
inherit EcoDataComponent. MainLayout becomes an EcoDataLayout and loses its
hand-wired LanguageChanged subscription and its Dispose override entirely; its
two [Event] handlers move to a new code-behind, which is what lets it change
base at all.

That last part is the constraint worth knowing: Tempest's razor frontend
matches @inherits by simple name, so it cannot see that EcoDataComponent
derives from StatefulComponent, and any [Command]/[Reactive]/[Event] written
inside an @code block under it raises TEM002. Declared in a .razor.cs partial
they resolve fine, because the C# symbol frontend walks the real base chain.
The base class documents this.

Which is why the seven pages keep StatefulComponent and their hand-wired
LanguageChanged for now: each carries three to six Tempest attributes inline,
and moving them all to code-behind is a much larger change than this one.
SpeciesGrid is also left alone — it implements IAsyncDisposable while the base
disposes synchronously, and that pair deserves its own look.

Verified the generator still emits for all fifteen Tempest components,
MainLayout included, now resolving through its code-behind.

dotnet build EcoData.slnx: 0 errors.
