+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Actions/ActionEditorialHero.razor
@@ -1,5 +1,5 @@
 @namespace FaunaFinder.Client.Components.Actions
-@inherits LocalizedComponentBase
+@inherits EcoDataComponent
 
 <section class="aa-hero">
     <div class="aa-hero-main">
