+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Map/MapConservationFilterDialog.razor
@@ -1,5 +1,5 @@
 @namespace FaunaFinder.Client.Components.Map
-@inherits LocalizedComponentBase
+@inherits EcoDataComponent
 
 <MudDialog Visible="Visible" VisibleChanged="VisibleChanged"
            Options="new DialogOptions { MaxWidth = MaxWidth.Small, FullWidth = true, CloseOnEscapeKey = true }">
