+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Map/MapNearbyPanel.razor
@@ -1,5 +1,5 @@
 @namespace FaunaFinder.Client.Components.Map
-@inherits LocalizedComponentBase
+@inherits EcoDataComponent
 @using EcoData.Wildlife.Contracts.Dtos
 
 <header class="mnp-head">
