+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Map/MapSpeciesPanel.razor
@@ -1,5 +1,5 @@
 @namespace FaunaFinder.Client.Components.Map
-@inherits LocalizedComponentBase
+@inherits EcoDataComponent
 @using EcoData.Wildlife.Contracts
 @using EcoData.Wildlife.Contracts.Dtos
 @using EcoData.Wildlife.Contracts.Parameters
