+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Municipalities/MunicipalityCard.razor
@@ -1,5 +1,5 @@
 @namespace FaunaFinder.Client.Components.Municipalities
-@inherits LocalizedComponentBase
+@inherits EcoDataComponent
 @using EcoData.Locations.Contracts.Dtos
 @using EcoData.Wildlife.Application.Client
 @using EcoData.Wildlife.Contracts.Parameters
