+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Municipalities/MunicipalityDetailCard.razor
@@ -1,5 +1,5 @@
 @namespace FaunaFinder.Client.Components.Municipalities
-@inherits LocalizedComponentBase
+@inherits EcoDataComponent
 @using EcoData.Locations.Contracts.Dtos
 
 <aside class="mm-detail">
