+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Municipalities/MunicipalityMap.razor
@@ -1,5 +1,5 @@
 @namespace FaunaFinder.Client.Components.Municipalities
-@inherits LocalizedComponentBase
+@inherits EcoDataComponent
 @using System.Text.Json
 @using EcoData.Spa.Map
 @using EcoData.Locations.Contracts.Dtos
