+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Municipalities/MunicipalityStatsRow.razor
@@ -1,5 +1,5 @@
 @namespace FaunaFinder.Client.Components.Municipalities
-@inherits LocalizedComponentBase
+@inherits EcoDataComponent
 @using EcoData.Wildlife.Contracts.Dtos
 
 <section class="mm-stats">
