+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Practices/PracticeEditorialHero.razor
@@ -1,5 +1,5 @@
 @namespace FaunaFinder.Client.Components.Practices
-@inherits LocalizedComponentBase
+@inherits EcoDataComponent
 
 <section class="pp-hero">
     <div class="pp-hero-main">
