+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Shell/FfBottomNav.razor
@@ -1,5 +1,5 @@
 @namespace FaunaFinder.Client.Components.Shell
-@inherits LocalizedComponentBase
+@inherits EcoDataComponent
 @inject INavigationManager Navigation
 
 @* The mobile tab bar — a floating pill, not a docked strip.
