+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Shell/FfBottomNav.razor.cs
@@ -1,3 +1,4 @@
+using EcoData.Spa.Blazor;
 using EcoData.Spa.Navigation;
 using EcoData.Spa.Navigation.Events;
 using FaunaFinder.Client.Localization;
@@ -10,9 +11,9 @@ namespace FaunaFinder.Client.Components.Shell;
 
 // The [Event] handlers live in this code-behind (with the base stated
 // explicitly) because Tempest's razor frontend matches the @inherits text by
-// simple name and can't see that LocalizedComponentBase is a StatefulComponent;
+// simple name and can't see that EcoDataComponent is a StatefulComponent;
 // the C# symbol frontend can.
-public partial class FfBottomNav : LocalizedComponentBase
+public partial class FfBottomNav : EcoDataComponent
 {
     /// <summary>
     /// Get the tab bar out of the way — published by a screen that needs the
