+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Shell/FfLibraryRail.razor
@@ -1,5 +1,5 @@
 @namespace FaunaFinder.Client.Components.Shell
-@inherits LocalizedComponentBase
+@inherits EcoDataComponent
 @using EcoData.Wildlife.Contracts.Dtos
 @using FaunaFinder.Client.Services.FieldNotebook
 @inject ISpeciesHttpClient SpeciesClient
