+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Shell/FfMobileTitle.razor
@@ -1,5 +1,5 @@
 @namespace FaunaFinder.Client.Components.Shell
-@inherits LocalizedComponentBase
+@inherits EcoDataComponent
 @inject INavigationManager Navigation
 @inject INavbarManager Navbar
