+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Shell/FfSearch.razor
@@ -1,5 +1,5 @@
 @namespace FaunaFinder.Client.Components.Shell
-@inherits LocalizedComponentBase
+@inherits EcoDataComponent
 @using EcoData.Locations.Contracts.Dtos
 @using EcoData.Wildlife.Contracts.Dtos
 @inject ISpeciesHttpClient SpeciesClient
