+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Species/SpeciesEditorialHero.razor
@@ -1,5 +1,5 @@
 @namespace FaunaFinder.Client.Components.Species
-@inherits LocalizedComponentBase
+@inherits EcoDataComponent
 @using EcoData.Wildlife.Contracts.Dtos
 
 <section class="ff-hero">
