+++ src/Apps/FaunaFinder/FaunaFinder.Client/Components/Species/SpeciesFeaturedRow.razor
@@ -1,5 +1,5 @@
 @namespace FaunaFinder.Client.Components.Species
-@inherits LocalizedComponentBase
+@inherits EcoDataComponent
 @using EcoData.Wildlife.Contracts.Dtos
 @using EcoData.Wildlife.Application.Client
 @inject ISpeciesHttpClient SpeciesClient
