+++ src/Apps/FaunaFinder/FaunaFinder.Client/_Imports.razor
@@ -21,6 +21,7 @@
 @using FaunaFinder.Client.Components.Actions
 @using FaunaFinder.Client.Components.Map
 @using FaunaFinder.Client.Components.Shared
+@using EcoData.Spa.Blazor
 @using FaunaFinder.Client.Localization
 @using EcoData.Common.i18n
 @using EcoData.NativeUi
